Extend the expiry of `Security.DataDecoder.Image...` histograms.
The histograms measure the overhead of using C++ (and a separate
utility process) to decode images.  We want to keep the histograms
until we are ready to consider a project that replaces the C++
code with 3rd-party Rust crates (not yet, we are still working
toward shipping first bits of Rust code in Chromium).

Fixed: 1424002
Fixed: 1424003
Change-Id: Ie42ccd0ccbff27c1bce156dfa5f389f79e4d79e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4461398
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1137982}