QA Rich Card: update icons on QuickAnswersView
As part of the Quick Answers rich card redesign, the Google icon on the
original Quick Answers card view will be replaced by the icons that
correspond to the result type. Add the icon changes behind the feature
flag, including the removal of the settings button (the settings button
will be migrated to the rich card).

Screenshots:
https://screenshot.googleplex.com/8inRpKDK54BooZS.png
https://screenshot.googleplex.com/9kwRpcbA7AJJKfo.png
https://screenshot.googleplex.com/BETzPPhTE9EGQDT.png
https://screenshot.googleplex.com/ArrD5GXKHgiaoRh.png
https://screenshot.googleplex.com/4tfWJPYA39iLCPQ.png
https://screenshot.googleplex.com/8jtwn6FeQNTpEvf.png

Bug: b/279614427
Change-Id: I813f12a5903e78d43c7ebb6742f922d7cc56a866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4485317
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Angela Xiao <angelaxiao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138006}