Refactor Browsing Topics to use a locally held ML Annotator
The PageContentAnnotationsService was meant to reused across many
features, but the BrowsingTopics model isn't going to be reused by
other features so this CL copies it into browsing_topics owned code
instead.

This CL is a code copy, intentionally not a move. Completing the code
removal from the Optimization Guide code will be much more work so
that will occur separately. The only code moved is a golden data
browser test which, if not moved, would otherwise cause DCHECK failures.

An added benefit of this refactor is that it will be much easier to
complete future changes to what input is given to the ML model.

Bug: b/278162907
Change-Id: I1bb57be46a6efc4ac690ba36675bf9c990c6081c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4476827
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138025}