[Read Anything] Pass AXTreeID to screen2x callback.
We should never pass a reference to a callback without any lifetime
tracking. Passing an AXTreeID removes this risk.

This also removes the call to distill by algorithm after a failed
screen2x attempt. We had discussed removing this before so this is an
opportune time.

Fixed: 1434669
Bug: 1266555
Change-Id: If03255c1ab0c4636e2b0580389850b1e5da21c53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4494011
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Kristi Saney <kristislee@google.com>
Cr-Commit-Position: refs/heads/main@{#1138051}