App Management: Switch permission toggles for ARC-T to read only
This replaces the existing permission items for ARC-T with a new "read
only permission item" component which is just an icon with labels
describing whether the permission is allowed/denied. This is a first
pass with minimal styling work, it's possible we'll want to tweak things
in response to UX feedback. Screenshot of UI on bug or at
http://screen/5zgTeQFqsBeS8JT.

Note that the ARC detail view now uses a dom-repeat to define the
permissions list, to void a whole lot of code duplication. In the future
we could consider refactoring this out into a 'permission list'
component for use in other app types.

Bug: 1433195
Change-Id: Iea9a447a367c8d89b8a921795ec6123ff3df3c25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4455913
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138064}