shortcuts: Limit search query length to 200 characters
This CL limits the max query length of a search to 200 characters using
the `maxlength` HTML attribute.

The primary purpose of this CL is to prevent performance degradation
when many lines of text are entered (usually pasted) into the search
bar.

As part of this CL, I created a new member variable called
`searchInputElement` that is queried for + saved on init. I've updated
`strictQueries` for this element to instead use the member variable.
This is primarily a cosmetic refactor, but it's possible that there
may be very slight performance gains as a result of this change.

I've also added a test for this change, checking that the maxLength
attribute has been set. We're limited in what we can test because we
can't simulate keypresses, and manually setting the value of the input
field doesn't respect the `maxlength` attribute.

Test: browser_tests gtest_filter=ShortcutCustomizationAppSearch*
Bug: b:266085562
Change-Id: I29ea66d28302e3a6a84abfcf398b21a50c0568d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4492077
Reviewed-by: Michael Checo <michaelcheco@google.com>
Commit-Queue: Camden Bickel <cambickel@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138067}