[CrOS Cellular] Matching last connected attach APN but not default.
Implement ApnMigrator case where the network is non-managed and its APN
matches the last connected attach APN but not the last connected
default APN.

If the last connected default APN matches with an APN in
the network's apn_list, the custom APN was used to attach to a database
default APN. In this circumstance, both the last connected attach
and default APNs will be migrated with Enabled state.

If the last connected default APN does not match with an APN in the
network's apn_list, the last connected attach APN will be migrated with
Disabled state.

Bug: b:162365553
Test: *ApnMigrator*
Change-Id: Ic50fde0ae5263a0c3b9e728ba3d2c68dbfff9d33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4483508
Reviewed-by: Gordon Seto <gordonseto@google.com>
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138103}