Assign DelayBasedTimeSource::timer_ task runner when it is activated
The DelayBasedTimeSource may be constructed on a different thread
than the one it runs on. For example,
blink::WidgetBase::RequestNewLayerTreeFrameSink() runs on the blink
main thread, and if --disable-frame-rate-limit is specified it will
construct a BackToBackBeginFrameSource synchronously. This code path
isn't currently used, but we plan to use it when we start running
web_tests with threaded compositing by default.

DelayBasedTimeSource::SetActive, which must run on the same thread as
task_runner_, is the first point where timer_ needs to be correctly
set up, so call SetTaskRunner there.

Bug: chromium:770028
Change-Id: Ie6e9bb472c645f952fa3b2f8fc3192ec3d4c8663
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4496884
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138127}