Use kAdjustmentResizeX for submenus
in some cases, the browser>history>specific window submenu is wide
enough that the compositor would fail to flip it in the X direction,
causing it to overflow out of the screen completely, resulting in a
suboptimal user experience. Fix this by letting the compositor fall back
to resizing submenus if flipping failed.

Bug: 1440960
Change-Id: Iee0a7f9165db98b0b468b7997a263737477a8a69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4487990
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138180}