[FilesBulkPinning] Expose kOkWithMoreResults drivefs result
When searching for items, it's possible to hit the end of the locally
available results but have some cloud results that are still potentially
available. This is only really required when the QuerySource is
`kLocalAndCloud` as it tries a local search first then defers to the
cloud.

Currently the logic is an empty page indicates the end of a query.
However when using `kLocalAndCloud` this logic is incorrect. To avoid
terminating the query at empty pages, provide another `FileError` called
`kOkWithMoreResults` to indicate the local results have been exhausted
and to try `GetNextPage` again to search the cloud. If the subsequent
search returns `kOk` with an empty page, this is the final result.

Bug: b:280240617, b:277018122
Test: TBD
Change-Id: I1d435a814b8f87731f606a0b4b6fe76eebcc53b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4495609
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138280}