[iOS] Make subclasses implement isItemAtIndexPathTextEditCell
isItemAtIndexPathTextEditCell from the protocol
AutofillEditTableViewController is used by passwords, credit card and
address settings to move focus forth and back to the fields in the
settings view.
By making the subclasses implement this method, whenever a new item
is introduced in the view, the subclass need to specify if the field
is a text field or not. This would prevent runtime crashes if new fields
are not text edit fields.

Fixed: 1429482
Change-Id: I4c1b5e88d07a28460b90992a621ed793edb0ac5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4490103
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Cr-Commit-Position: refs/heads/main@{#1138281}