iOS sync settings: Ensure IsSetupInProgress is true while settings open
While any Sync-settings UI is open, it should hold a SyncSetupInProgress
handle, to ensure that SyncService::IsSetupInProgress() is true as long
as the UI is open. (This ensures that any pending changes don't get
applied until the user confirms/closes the UI.)

Previously, the iOS sync settings UI did this, but a bit inconsistently
(basically it grabbed a SyncSetupInProgressHandle on the first change,
but (a) it's not clear if that actually covered all changes, and (b)
with some upcoming changes that's not good enough, since
IsSetupInProgress() needs to be true as soon as the UI is opened).

After this CL, the settings coordinator explicitly marks setup as in
progress before showing the UI. The previous ad-hoc logic in
SyncSetupService is replaced by CHECKs.
This makes the IsSetupIsProgress() state more reliable, and consistent
with how it works on all other platforms.

Bug: 1430453, 1438800
Change-Id: I1084d045a17376528eed7a9e695088fd68368f63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4487872
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138287}