[iOS][sync] Avoid IsSyncRequested() in ManageSyncSettingsMediator
The usage of IsSyncRequested() to gray out the toggles appears to be
legacy, from the time when the UI allowed disabling the sync-requested
bit.

Starting with https://crrev.com/3625480, IsSyncRequested() is always
true, except for one edge case, which is the only behavioral change in
this patch: if the user signs out, there is some short time during
which the UI continues to be displayed. Before this patch, the toggles
turned gray during this time, whereas after this patch nothing changes
in the UI (which is arguably a better UX). In the hypothetical case
where the user managed to interact with the toggles in this short time
interval, nothing really bad happens (it's similar to interacting with
the toggles *before* signing out).

This removes the last remaining calling site to the deprecated
SyncSetupService::IsSyncRequested().

Bug: 1291953
Change-Id: I6efed7d3f60bbfe12219e84b3111bf37dca972c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4487856
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138358}