Add metrics for whether the memory savings page action chip is expanded
This histogram will be logged only once for each discarded page that is
revisited. It will capture whether the chip was rendered collapsed, or
expanded due to it being the first time a user saw it or to highlight
memory savings.

This also ensures that the SetWasAnimated flag is set immediately once entering that loop in UpdateImpl so the collapsed histogram isn't logged multiple times.

Bug: 1424230
Change-Id: I3415d3e0b284d8866afa4baed6f82bbbc20ee63f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4496785
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Commit-Queue: Alison Gale <agale@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138369}