[iOS] Fix compromised state of password is not updated after an edit
This CL rolls back a change made in the PasswordDetailsMediator (see
link below) to again use the list of insecure credentials provided by
the IOSChromePasswordCheckManager to determine whether a credential is
compromised or not. The list of insecure credentials was indeed
necessary because the `_credentials` list contained in the
PasswordDetailsMediator is not updated when a credential is edited in
the details page. Because of that, the updates on a credential were not
always correctly reflected in the details page. This change also removes
the code that was locally muting a compromised password since the
insecure credentials list can now be used instead.

Link of the change made in PasswordDetailsMediator:
https://chromium-review.googlesource.com/c/chromium/src/+/4415224/6/ios/chrome/browser/ui/settings/password/password_details/password_details_mediator.mm#b374

Fixed: 1440794
Change-Id: I6b377aa3e35c3ee3c9b3f096b37a452c0709b79b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4491726
Auto-Submit: Noémie St-Onge <noemies@google.com>
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Commit-Queue: Veronique Nguyen <veronguyen@google.com>
Reviewed-by: Veronique Nguyen <veronguyen@google.com>
Cr-Commit-Position: refs/heads/main@{#1138396}