Remove SSLVersionMin Policy
This policy has been obsolete since v98 and is no longer respected as
the only supported token, |tls1.2|, is also the default. The majority of
the implementation code remains behind to support the |ssl-version-min|
command line argument which is still respected and allows a |tls1.3|
token.

Bug: 1155619
Change-Id: I24c7278101682bc12b32e5b751329ae280fc4839
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4494702
Commit-Queue: Eric Lawrence <ericlaw@microsoft.com>
Reviewed-by: Sergiy Belozorov <sergiyb@google.com>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138429}