Add FeatureParam for memory threshold percentiles
These will be used for determining the breakpoints for the gauge
style chart inside the memory saver dialog. There will be 4 buckets
in the chart, so we need 3 percentile values to place savings into
the appropriate bucket. Data comes from a UKM Colab and we want it
to be Finchable in case we need to adjust the buckets.

Bug: 1424230
Change-Id: I7d5bdec303d0caf8f4dffd74744cbd68cc33d386
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4496869
Commit-Queue: Alison Gale <agale@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1138434}