[iOS] Enable account storage for bookmark and readinglist when sign-in
The account storage is enabled when AuthenticationFlow is used with
kEnableBookmarkReadingListAccountStorage post action.
This is done with the sign-in promo in bookmark or reading list
sign-in promo.

kEnableBookmarksAccountStorage flag or
(kReadingListEnableDualReadingListModel and
kReadingListEnableSyncTransportModeUponSignIn) flags have to be enabled.

Bug: 1427044
Change-Id: I66766241534e3311853d02ed9680620a525b17ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4498392
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138448}