[Android] Bottom sheet hides on scroll also in half state
This CL updates BottomSheet to allow the behavior 'hide-on-scroll' in
half state as well as peek state.

The change does not affect any existing bottom sheets since there are
none that enables the behavior. If a new one enabled it, the scrim
would ordinarily prevent it past the peeking state.

Change-Id: I494037c8756873cbf80c09ff3e7addd15d2e9457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4494671
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1138460}