[QuickStart][1/n] Add QuickStartDecoder to TargetDeviceConnectionBroker's Constructor
This is part of a series of changes to allow connections to be initialized with the
decoder.

This change adds QuickStartDecoder as a constructor argument for
TargetDeviceConnectionBroker. It is added as a raw dependency (rather than through
the NearbyProcessReference) to better enable dependency injection for
setup.

Future changes will modify both ends of this pipe to thread all the way
from QuickStartConnectivityService

Test: Updated unit tests
Bug: b/280330703
Change-Id: I0fea3fd8901a2a916e2609ca90d375bf01c5e2b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4493711
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Commit-Queue: Joe Antonetti <joeantonetti@google.com>
Cr-Commit-Position: refs/heads/main@{#1138471}