lacros/color: preload common colorspaces
The Chrome Color Management protocol currently needs a roundtrip between
the client and server the first time a colorspace is requested for use.
This means for certain content the first frame will be in a wrong or
default colorspace before switching over to the correct one. This can
cause flashing.

This CL mitigates the issue some by preloading the colorspaces commonly
seen on the web (youtube, google meets, hdr content) so they are ready
the first time they are requested.

BUG=1429074, b:262590827
TEST=build and deploy to dut

Change-Id: I5d9202778b6ce2bda3f43dffe24b1a2c6bd31a1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4497562
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Commit-Queue: Femi Adegunloye <mrfemi@google.com>
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138473}