Create exception via high efficiency dialog bubble
Add another button to the high efficiency dialog bubble that will give
users the option to add the discarded site to the exceptions list.

Re-opening the dialog bubble after adding the site to the exceptions
list will show a disabled button instead because the site was already
added to the exceptions list.

Bug: 1424224
Change-Id: I32a8d98cee5e485c8d3d39d7f7c4d592464ef111
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4443309
Commit-Queue: Steven Luong <stluong@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138475}