Add controller type check to IsLocalMirroringRoute
Right now, IsLocalMirroringRoute on checks if the source is for a tab
mirroring session or a desktop mirroring session. However, this misses
site-initiated mirroring sessions. For example, when Google slides
starts casting. In all cases where CastActivityManager will create an
instance of MirroringActivity for a session, it sets the route's
controller_type to RouteControllerType::kMirroring. So, we can use this
to check whether the route has an associated MirroringActivity.

Bug: b:272368609
Change-Id: Ifd6a3d05d3f5572bae91f49d79d8ff7f33d923b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4493023
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Commit-Queue: Benjamin Zielinski <bzielinski@google.com>
Cr-Commit-Position: refs/heads/main@{#1138514}