Allow same-site prefetches to be made in the default network context.
Previously only same-origin URLs could be prefetched in the default
network context. This CL changes it so same-site prefetches are made in
the default network context, and cross-site prefetches are made in
isolated network contexts.

The narrow case of same-site cross-origin prefetches that require the
prefetch proxy are blocked, because it would require configuring the
default network context to use the proxy for just those prefetches. This
is tracked in a separate bug crbug.com/1439986.

Bug: 1414582
Change-Id: Ic1bcda6c1a253f668315dfe7ad2ad623a9eebabb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4482043
Reviewed-by: Simon Pelchat <spelchat@chromium.org>
Commit-Queue: Max Curran <curranmax@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1138520}