[QuickStart][3/n] Refactor TargetDeviceBootstrapController to no longer call factory directly
This is part of a series of changes to pass QuickStartDecoder down to
the Connection object directly.

This change refactors TargetDeviceBootstrapController to accept a
TargetDeviceConnectionBroker, rather than constructing it directly using
a factory. This change was made for several reasons:

1) Since TargetDeviceBootstrapController only uses one instance of the
connection broker created in it's constructor, it has no need to have
a reference to a factory (it never creates arbitrary instances)
2) As we increase the number of dependencies needed to construct
TargetDeviceConnectionBroker, threading all of these dependencies through
TargetDeviceBootstrapController will become difficult to manage

Test: Updated unit tests
Bug: b/280330703
Change-Id: Ib232c0ee290f00a3af64384908f7aa75a2ba983c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4495511
Commit-Queue: Joe Antonetti <joeantonetti@google.com>
Reviewed-by: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/main@{#1138525}